Work at: - Fix notification bug; - Display schedule in Proposal details; - Select Proposal to send mail from index / details; - Proposal:   . Reject option; - Redirect to home when approve / reject proposal; - CTPS:   . State => FINISHED; - Bug fixes;

+++ database/migrations/2014_10_12_000001_create_users_table.php
@@ -13,19 +13,21 @@ class CreateUsersTable extends Migration
      */
     public function up()
     {
-        Schema::create('users', function (Blueprint $table) {
-            $table->bigIncrements('id');
+        if (config('broker.useSSO')) {
+            Schema::create('users', function (Blueprint $table) {
+                $table->bigIncrements('id');
 
-            $table->string('name');
-            $table->string('email')->unique();
-            $table->timestamp('email_verified_at')->nullable();
-            $table->string('phone', 11)->nullable(true)->default(null);
-            $table->string('password');
-            $table->string('api_token', 80)->unique()->nullable()->default(null);
+                $table->string('name');
+                $table->string('email')->unique();
+                $table->timestamp('email_verified_at')->nullable();
+                $table->string('phone', 11)->nullable(true)->default(null);
+                $table->string('password');
+                $table->string('api_token', 80)->unique()->nullable()->default(null);
 
-            $table->rememberToken();
-            $table->timestamps();
-        });
+                $table->rememberToken();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -35,6 +37,8 @@ class CreateUsersTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('users');
+        if (config('broker.useSSO')) {
+            Schema::dropIfExists('users');
+        }
     }
 }
